Final Program

+++ stu_a0003/asg_a0006/AsgA0006.java
@@ -1,5 +1,5 @@
 public class AsgA0006 {
-public static double get_amount(double a,int b,int c)
+public static double get_amount(double a,double b,double c)
 {
 double simple_intrest=(a*b*c)/100;
         return(simple_intrest);
